feat(shared): M7 finish — TMDB + Plugin wire types migrated
packages/shared picks up the second + third audit targets for the shared
workspace. Only wire contracts (types that cross the HTTP boundary
between backend and frontend) move — runtime-side-only types that
reference Fastify, Prisma, or DOM stay where they are.

TMDB — new shared/src/tmdb.ts with: TmdbSeason, TmdbCast, TmdbCrew,
TmdbVideo, TmdbCollectionRef, TmdbCollection, TmdbMovie, TmdbTv,
TmdbMedia, TmdbMediaResult, TmdbPerson. TmdbPerson's combined_credits
now uses the loose TmdbMedia shape — the backend's previous
'TmdbMovie & TmdbTv & …' was lying (those entries are actually mixed
movie+tv rows with partial data). Call sites that were relying on
.title AND .name existing at the same time already handled the
fallback — the stricter-looking type never caught a real bug.

Plugins — new shared/src/plugins.ts with: PluginUIContribution,
PluginInfo, PluginSettingDef, PluginSettings. Kept out:
PluginManifest + UIContribution + PluginContext stay in backend
(they reference FastifyBaseLogger, NotificationRegistry, ArrClient,
and other runtime types that have no business in a pure-types package).

Call-site impact — backend services/tmdb.ts + plugins/types.ts and
frontend types/index.ts + plugins/types.ts now re-export from
@oscarr/shared. Existing import paths (@/types, services/tmdb.js,
plugins/types.js) keep working via re-export — follow-up commits can
migrate individual sites to direct @oscarr/shared imports at leisure.
No functional changes, typecheck green on all three packages.
