+++ packages/shared/src/index.ts
@@ -1,1 +1,3 @@
 export * from './requestStatus.js';
+export * from './tmdb.js';
+export * from './plugins.js';
